requirement.txt
absl-py==0.7.1
aiohttp==3.5.4
altgraph==0.16.1
astor==0.7.1
async-timeout==3.0.1
atari-py==0.2.6
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cloudpickle==1.2.1
cycler==0.10.0
Cython==0.29.7
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
future==0.17.1
gast==0.2.2
glfw==1.8.2
google-pasta==0.1.5
graphviz==0.8.4
grpcio==1.20.1
gym==0.13.1
h5py==2.9.0
idna==2.8
idna-ssl==1.1.0
imageio==2.5.0
ipykernel==5.1.1
ipython==7.6.0
ipython-genutils==0.2.0
ipywidgets==7.5.0
jedi==0.14.0
Jinja2==2.10.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.5.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
lmdb==0.94
lockfile==0.12.2
macholib==1.11
Markdown==3.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mistune==0.8.4
mock==2.0.0
multidict==4.5.2
mxnet-cu100==1.5.0b20190703
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
numpy==1.14.6
opencv-python==4.1.0.25
pandas==0.25.0
pandocfilters==1.4.2
parso==0.5.0
pbr==5.1.3
pefile==2019.4.18
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
pkg-resources==0.0.0
prettytable==0.7.2
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.7.1
ptyprocess==0.6.0
pycparser==2.19
pyecharts==1.1.0
pyglet==1.3.2
Pygments==2.4.2
PyInstaller==3.4
pynput==1.4.2
pyparsing==2.4.0
pyrsistent==0.15.2
python-dateutil==2.8.0
python-xlib==0.25
pyzmq==18.0.1
qtconsole==4.5.1
redis==3.3.5
requests==2.22.0
scipy==1.3.0
Send2Trash==1.5.0
six==1.12.0
tb-nightly==1.14.0a20190301
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tf-estimator-nightly==1.14.0.dev2019030115
torch==1.0.1.post2
torchvision==0.2.2.post3
tornado==6.0.3
tqdm==4.32.2
traitlets==4.3.2
typing-extensions==3.7.2
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.2
widgetsnbextension==3.5.0
xlrd==1.2.0
yarl==1.3.0
zmq==0.0.0
